fix: modify dir structure
